Initial Commit/Start of Unit 0 Proj 1
What I have done is using the source code from the lecture I am modifying so that it can work within this project parameters. I am having an issue but I think it has to with how the node is being created and finding out what the possible paths are. I plan to draw out what it is doing now so I can solve the error.