Print Statement Debug
What I did during this session was focus on understanding what data and how it is being passed. I did this using print statements. I drew out a map following the nodes. From there I realized that it was finding a path but something was not passing through. I though first that the datatype was not matching in the found true or false statement. However, it was because it was not returning the path. I allowed that and now running into a Type Error. I want to redo this for loop next time so it will be easier to navigate by using a key, value for loop.