Making it  function
I started over the error. I used the source code given and realized the error was in the tuple. I printed the tuple and realized all they had to was flip the index's. The plan is to max it out by using the A* option and turn it in tomorrow.