Final Submit
I was able to follow the notes from the lecture and was much familiar with the cs50 systems. That helped me find my errors. The issue I ran to was in the minmax function. I originally did not have the for loop based on the actions so it tried a value as a cell. That would through an error. My process for debugging was working backwards. I used print statements to figure out that it was inputting a "0". I just looked where that was an input. That was at best_action. Made them actual action and worked from there.